Phong Shading, Materials, Screenshots, C++17
-Implemented C++17
-Added Blinn-Phong shading
-Added textures
-Added option to switch between shaders
-Added screenshot function
-Added materials to work with shaders

+++ 02-Materials/include/Shapes.h
@@ -30,7 +30,7 @@ protected:
   public:
     std::vector<unsigned int> indices;
     std::vector<float> vertices;
-    float triangles = 0;
+    int triangles = 0;
     unsigned int *GetIndices();
     float *GetVertices();
     // Generates a sphere vertices containing position and normal
